Upstream version 5.1.0-dfsg

+++ src/VBox/ValidationKit/testdriver/vbox.py
@@ -27,7 +27,7 @@ CDDL are applicable instead of those of the GPL.
 You may elect to license modified versions of this file under the
 terms and conditions of either the GPL or the CDDL or both.
 """
-__version__ = "$Revision: 108087 $"
+__version__ = "$Revision: 108584 $"
 
 
 # Standard Python imports.
@@ -225,7 +225,7 @@ class ComError(object):
             if sAttr[0].isupper():
                 oAttr = getattr(oNativeComErrorClass, sAttr);
                 setattr(ComError, sAttr, oAttr);
-                if type(oAttr) is int:
+                if isinstance(oAttr, int):
                     ComError.dDecimalToConst[oAttr] = sAttr;
 
         # Install mappings to the other platform.
@@ -651,7 +651,6 @@ class ConsoleEventHandlerBase(EventHandlerBase):
     def onShowWindow(self):
         reporter.log2('onShowWindow/%s' % (self.sName));
         return None;
-
     # pylint: enable=C0111,R0913,W0613
 
     def handleEvent(self, oEvt):
@@ -754,6 +753,19 @@ class SessionConsoleEventHandler(ConsoleEventHandlerBase):
 
     def onMachineStateChange(self, sMachineId, eState):                         # pylint: disable=W0613
         """ Just interrupt the wait loop here so it can check again. """
+        _ = sMachineId; _ = eState;
+        self.oVBoxMgr.interruptWaitEvents();
+
+    def onRuntimeError(self, fFatal, sErrId, sMessage):
+        reporter.log('onRuntimeError/%s: fFatal=%d sErrId=%s sMessage=%s' % (self.sName, fFatal, sErrId, sMessage));
+        oSession = self.oSession;
+        if oSession is not None: # paranoia
+            if sErrId == 'HostMemoryLow':
+                oSession.signalHostMemoryLow();
+                if sys.platform == 'win32':
+                    from testdriver import winbase;
+                    winbase.logMemoryStats();
+            oSession.signalTask();
         self.oVBoxMgr.interruptWaitEvents();
 
 
@@ -837,8 +849,8 @@ class TestDriver(base.TestDriver):                                              
         if self.oBuild is not None:
             return True;
 
-        # Try dev build first since that's where I'll be using it first..
-        if True:
+        # Try dev build first since that's where I'll be using it first...
+        if True is True:
             try:
                 self.oBuild = Build(self, None);
                 return True;
@@ -909,7 +921,7 @@ class TestDriver(base.TestDriver):                                              
         # Do the searching.
         #
         sCandidate = None;
-        for i in range(len(asCandidates)):
+        for i, _ in enumerate(asCandidates):
             sCandidate = asCandidates[i];
             if os.path.isfile(os.path.join(sCandidate, 'VBoxValidationKit.iso')):
                 break;
@@ -1553,9 +1565,10 @@ class TestDriver(base.TestDriver):                                              
             if iArg >= len(asArgs):
                 raise base.InvalidOption('The "--vrdp-base-port" takes an argument');
             try:    self.uVrdpBasePort = int(asArgs[iArg]);
-            except: raise base.InvalidOption('The "--vrdp-base-port" value "%s" is not a valid integer', asArgs[iArg]);
+            except: raise base.InvalidOption('The "--vrdp-base-port" value "%s" is not a valid integer' % (asArgs[iArg],));
             if self.uVrdpBasePort <= 0 or self.uVrdpBasePort >= 65530:
-                raise base.InvalidOption('The "--vrdp-base-port" value "%s" is not in the valid range (1..65530)', asArgs[iArg]);
+                raise base.InvalidOption('The "--vrdp-base-port" value "%s" is not in the valid range (1..65530)'
+                                         % (asArgs[iArg],));
         elif asArgs[iArg] == '--vbox-default-bridged-nic':
             iArg += 1;
             if iArg >= len(asArgs):
@@ -1764,6 +1777,7 @@ class TestDriver(base.TestDriver):                                              
         try:
             rc = self.oVBoxMgr.waitForEvents(int(cMsTimeout));
             _ = rc; #reporter.log2('vbox.waitForTasksSleepWorker(%u): true (waitForEvents -> %s)' % (cMsTimeout, rc));
+            reporter.doPollWork('vbox.TestDriver.waitForTasksSleepWorker');
             return True;
         except KeyboardInterrupt:
             raise;
@@ -2275,6 +2289,7 @@ class TestDriver(base.TestDriver):                                              
                     reporter.errorTimeout('Timeout while waiting for progress.')
                 return -1
 
+            reporter.doPollWork('vbox.TestDriver.waitOnProgress');
             try:    oProgress.waitForCompletion(cMsInterval);
             except: return -2;
 
@@ -2334,13 +2349,13 @@ class TestDriver(base.TestDriver):                                              
         """
         fRc = True;
         try:
-            sLogFile = os.path.join(oVM.logFolder, 'VBoxStartup.log');
+            sLogFile = os.path.join(oVM.logFolder, 'VBoxHardening.log');
         except:
             reporter.logXcpt();
             fRc = False;
         else:
             if os.path.isfile(sLogFile):
-                reporter.addLogFile(sLogFile, 'log/release/vm', '%s startup log' % (sVmName, ),
+                reporter.addLogFile(sLogFile, 'log/release/vm', '%s hardening log' % (sVmName, ),
                                     sAltName = '%s-%s' % (sVmName, os.path.basename(sLogFile),));
         return fRc;
 
@@ -2542,63 +2557,83 @@ class TestDriver(base.TestDriver):                                              
         oSession.getPid();
 
         #
-        # Take Screenshot and upload it (see below) to Test Manager if appropriate/requested.
+        # If the host is out of memory, just skip all the info collection as it
+        # requires memory too and seems to wedge.
         #
         sLastScreenshotPath = None;
-        if fTakeScreenshot is True  or  self.fAlwaysUploadScreenshots  or  reporter.testErrorCount() > 0:
-            sLastScreenshotPath = os.path.join(self.sScratchPath, "LastScreenshot-%s.png" % oSession.sName);
-            fRc = oSession.takeScreenshot(sLastScreenshotPath);
-            if fRc is not True:
-                sLastScreenshotPath = None;
+        sOsKernelLog        = None;
+        sVgaText            = None;
+        asMiscInfos         = [];
+        if not oSession.fHostMemoryLow:
+            #
+            # Pause the VM if we're going to take any screenshots or dig into the
+            # guest.  Failures are quitely ignored.
+            #
+            if self.fAlwaysUploadLogs or reporter.testErrorCount() > 0:
+                try:
+                    if oSession.oVM.state in [ vboxcon.MachineState_Running,
+                                               vboxcon.MachineState_LiveSnapshotting,
+                                               vboxcon.MachineState_Teleporting ]:
+                        oSession.o.console.pause();
+                except:
+                    reporter.logXcpt();
 
-        # Query the OS kernel log from the debugger if appropriate/requested.
-        sOsKernelLog = None;
-        if self.fAlwaysUploadLogs or reporter.testErrorCount() > 0:
-            sOsKernelLog = oSession.queryOsKernelLog();
+            #
+            # Take Screenshot and upload it (see below) to Test Manager if appropriate/requested.
+            #
+            if fTakeScreenshot is True  or  self.fAlwaysUploadScreenshots  or  reporter.testErrorCount() > 0:
+                sLastScreenshotPath = os.path.join(self.sScratchPath, "LastScreenshot-%s.png" % oSession.sName);
+                fRc = oSession.takeScreenshot(sLastScreenshotPath);
+                if fRc is not True:
+                    sLastScreenshotPath = None;
 
-        # Do "info vgatext all" separately.
-        sVgaText = None;
-        if self.fAlwaysUploadLogs or reporter.testErrorCount() > 0:
-            sVgaText = oSession.queryDbgInfoVgaText();
+            # Query the OS kernel log from the debugger if appropriate/requested.
+            if self.fAlwaysUploadLogs or reporter.testErrorCount() > 0:
+                sOsKernelLog = oSession.queryOsKernelLog();
 
-        # Various infos (do after kernel because of symbols).
-        asMiscInfos = [];
-        if self.fAlwaysUploadLogs or reporter.testErrorCount() > 0:
-            # Dump the guest stack for all CPUs.
-            cCpus = oSession.getCpuCount();
-            if cCpus > 1024: # disable for now
-                for iCpu in xrange(0, cCpus):
-                    sThis = oSession.queryDbgGuestStack(iCpu);
+            # Do "info vgatext all" separately.
+            if self.fAlwaysUploadLogs or reporter.testErrorCount() > 0:
+                sVgaText = oSession.queryDbgInfoVgaText();
+
+            # Various infos (do after kernel because of symbols).
+            if self.fAlwaysUploadLogs or reporter.testErrorCount() > 0:
+                # Dump the guest stack for all CPUs.
+                cCpus = oSession.getCpuCount();
+                if cCpus > 0:
+                    for iCpu in xrange(0, cCpus):
+                        sThis = oSession.queryDbgGuestStack(iCpu);
+                        if sThis is not None and len(sThis) > 0:
+                            asMiscInfos += [
+                                '================ start guest stack VCPU %s ================\n' % (iCpu,),
+                                sThis,
+                                '================ end guest stack VCPU %s ==================\n' % (iCpu,),
+                            ];
+
+                for sInfo, sArg in [ ('mode', 'all'),
+                                     ('fflags', ''),
+                                     ('cpumguest', 'verbose all'),
+                                     ('cpumguestinstr', 'symbol all'),
+                                     ('pic', ''),
+                                     ('apic', ''),
+                                     ('ioapic', ''),
+                                     ('pit', ''),
+                                     ('phys', ''),
+                                     ('clocks', ''),
+                                     ('timers', ''),
+                                     ('gdtguest', ''),
+                                     ('ldtguest', ''),
+                                    ]:
+                    if sInfo in ['apic',] and self.fpApiVer < 5.1: # asserts and burns
+                        continue;
+                    sThis = oSession.queryDbgInfo(sInfo, sArg);
                     if sThis is not None and len(sThis) > 0:
+                        if sThis[-1] != '\n':
+                            sThis += '\n';
                         asMiscInfos += [
-                            '================ start guest stack VCPU %s ================\n' % (iCpu,),
+                            '================ start %s %s ================\n' % (sInfo, sArg),
                             sThis,
-                            '================ end guest stack VCPU %s ==================\n' % (iCpu,),
+                            '================ end %s %s ==================\n' % (sInfo, sArg),
                         ];
-
-            for sInfo, sArg in [ ('mode', 'all'),
-                                 ('fflags', ''),
-                                 ('cpumguest', 'verbose all'),
-                                 ('cpumguestinstr', 'symbol all'),
-                                 ('pic', ''),
-                                 ('apic', ''),
-                                 ('ioapic', ''),
-                                 ('pit', ''),
-                                 ('phys', ''),
-                                 ('clocks', ''),
-                                 ('timers', ''),
-                                 ('guestgdt', ''),
-                                 ('ldtguest', ''),
-                                ]:
-                sThis = oSession.queryDbgInfo(sInfo, sArg);
-                if sThis is not None and len(sThis) > 0:
-                    if sThis[-1] != '\n':
-                        sThis += '\n';
-                    asMiscInfos += [
-                        '================ start %s %s ================\n' % (sInfo, sArg),
-                        sThis,
-                        '================ end %s %s ==================\n' % (sInfo, sArg),
-                    ];
 
         #
         # Terminate the VM
@@ -2618,34 +2653,35 @@ class TestDriver(base.TestDriver):                                              
 
         # Check if the VM has terminated by it self before powering it off.
         fClose = True;
-        fRc = oSession.pollTask();
-        if fRc is not True:
+        fRc    = True;
+        if oSession.needsPoweringOff():
             reporter.log('terminateVmBySession: powering off "%s"...' % (oSession.sName,));
             fRc = oSession.powerOff(fFudgeOnFailure = False);
-        if fRc is not True:
-            # power off failed, try terminate it in a nice manner.
-            fRc = False;
-            uPid = oSession.getPid();
-            if uPid is not None:
-                reporter.error('terminateVmBySession: Terminating PID %u (VM %s)' % (uPid, oSession.sName));
-                fClose = base.processTerminate(uPid);
-                if fClose is True:
-                    self.waitOnDirectSessionClose(oSession.oVM, 5000);
-                    fClose = oSession.waitForTask(1000);
-
-                if fClose is not True:
-                    # Being nice failed...
-                    reporter.error('terminateVmBySession: Termination failed, trying to kill PID %u (VM %s) instead' \
-                                   % (uPid, oSession.sName));
-                    fClose = base.processKill(uPid);
+            if fRc is not True:
+                # power off failed, try terminate it in a nice manner.
+                fRc = False;
+                uPid = oSession.getPid();
+                if uPid is not None:
+                    reporter.error('terminateVmBySession: Terminating PID %u (VM %s)' % (uPid, oSession.sName));
+                    fClose = base.processTerminate(uPid);
                     if fClose is True:
                         self.waitOnDirectSessionClose(oSession.oVM, 5000);
                         fClose = oSession.waitForTask(1000);
+
                     if fClose is not True:
-                        reporter.error('terminateVmBySession: Failed to kill PID %u (VM %s)' % (uPid, oSession.sName));
+                        # Being nice failed...
+                        reporter.error('terminateVmBySession: Termination failed, trying to kill PID %u (VM %s) instead' \
+                                       % (uPid, oSession.sName));
+                        fClose = base.processKill(uPid);
+                        if fClose is True:
+                            self.waitOnDirectSessionClose(oSession.oVM, 5000);
+                            fClose = oSession.waitForTask(1000);
+                        if fClose is not True:
+                            reporter.error('terminateVmBySession: Failed to kill PID %u (VM %s)' % (uPid, oSession.sName));
 
         # The final steps.
         if fClose is True:
+            reporter.log('terminateVmBySession: closing session "%s"...' % (oSession.sName,));
             oSession.close();
             self.waitOnDirectSessionClose(oSession.oVM, 10000);
             try:
@@ -3149,6 +3185,7 @@ class TestDriver(base.TestDriver):                                              
 
         # Start the VM.
         reporter.log('startVmAndConnectToTxsViaTcp: Starting(/preparing) "%s" (timeout %s s)...' % (sVmName, cMsTimeout / 1000));
+        reporter.flushall();
         oSession = self.startVmByName(sVmName);
         if oSession is not None:
             # Connect to TXS.
